update LuaGlue to the latest version
Former-commit-id: 00d0799eeb065d525fb9fa111988090639d664a2

+++ plugins/lua/lib/LuaGlue/include/LuaGlue/LuaGlueDebug.h
@@ -3,6 +3,7 @@
 
 #ifdef DEBUG
 
+#include <cstring>
 static inline void LG_Debug_(const char *FILE, const char *FUNCTION, int LINE, const char *t, const char *format, ...)
 {
 	va_list ap;
@@ -17,7 +18,7 @@ static inline void LG_Debug_(const char *FILE, const char *FUNCTION, int LINE, c
 	msg[msg_len] = 0;
 	
 	va_start(ap, format);
-	vsnprintf(msg, msg_len+1, format, ap);
+	vsnprintf(msg, (size_t)msg_len+1, format, ap);
 	va_end(ap);
 	
 	const char *fptr = strrchr(FILE, '/');
